+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Change log
 
-### 1.7.6 (4/28 - )
+### 1.7.6 (4/28 - 5/17)
 * ansible: jenkins supports plugins
 * provision: use private ip if public dns not available, for build server provision private subnet server (by Gabo)
+* deploy: retry on throttling of ASG update
 
 ### 1.7.5 (3/28/2016 - 4/28)
 * mongodb: clustering support